tarefa

+++ Projeto/tarefas/src/main/java/com/example/tarefas/service/tarefa/TarefaService.java
@@ -3,14 +3,19 @@ package com.example.tarefas.service.tarefa;
 
 import com.example.tarefas.controller.tarefa.dto.TarefaDto;
 import com.example.tarefas.controller.usuario.dto.UsuarioDto;
+import com.example.tarefas.exceptions.BadRequestException;
 import com.example.tarefas.model.Tarefa;
 import com.example.tarefas.model.Usuario;
 import com.example.tarefas.repository.TarefaRepository;
 import com.example.tarefas.repository.UsuarioRepository;
 import jakarta.persistence.EntityNotFoundException;
+import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Service;
+import org.springframework.web.multipart.MultipartFile;
+import java.io.BufferedReader;
+import java.io.InputStreamReader;
 
 
 import java.time.LocalDate;
@@ -19,11 +24,17 @@ import java.util.Optional;
 
 @Service
 public class TarefaService {
-    @Autowired
-    TarefaRepository tarefaRepository;
 
+    @Autowired
+    private TarefaRepository tarefaRepository;
+
+    @Autowired
+    private UsuarioRepository usuarioRepository;
+
+    @Transactional
     public List<Tarefa> findAll() { return tarefaRepository.findAll(); }
 
+    @Transactional
     public Tarefa findById(long id) {
         Optional<Tarefa> task  = tarefaRepository.findById(id);
         if(task.isPresent()) {
@@ -36,8 +47,7 @@ public class TarefaService {
 
 
 
-    @Autowired
-    private UsuarioRepository usuarioRepository;
+
 
     public Tarefa create(TarefaDto tarefaDto) {
         Usuario usuario = usuarioRepository.findById(tarefaDto.usuarioId())
@@ -53,8 +63,13 @@ public class TarefaService {
         return tarefaRepository.save(tarefa);
     }
 
-    public Tarefa update(Long id, TarefaDto tarefaDto) {
+    public Tarefa update(Long id, TarefaDto tarefaDto, Long usuarioId) {
         Tarefa tarefaExistente = findById(id);
+
+        if (!tarefaExistente.getUsuario_criado().getId().equals(tarefaDto.usuarioId())) {
+            throw new BadRequestException("Você não tem permissão para alterar esta tarefa.");
+        }
+
 
         Usuario usuario = usuarioRepository.findById(tarefaDto.usuarioId())
                 .orElseThrow(() -> new RuntimeException("Usuário não encontrado"));
@@ -69,16 +84,67 @@ public class TarefaService {
         return tarefaRepository.save(tarefaAtualizada);
     }
 
-    public Tarefa alternarConclusao(Long id) {
+    public Tarefa concluirTarefa(Long id, Long usuarioId) {
         Tarefa tarefa = findById(id);
-        tarefa.setConcluido(!Boolean.TRUE.equals(tarefa.getConcluido()));
+
+        if (!tarefa.getUsuario_criado().getId().equals(usuarioId)) {
+            throw new BadRequestException("Você não tem permissão para concluir esta tarefa.");
+        }
+
+        tarefa.setConcluido(true);
+        return tarefaRepository.save(tarefa);
+    }
+
+    public Tarefa pendenteTarefa(Long id, Long usuarioId) {
+        Tarefa tarefa = findById(id);
+
+        if (!tarefa.getUsuario_criado().getId().equals(usuarioId)) {
+            throw new BadRequestException("Você não tem permissão para alterar esta tarefa.");
+        }
+
+        tarefa.setConcluido(false);
         return tarefaRepository.save(tarefa);
     }
 
 
-    public void delete(Long id) {
-        findById(id);
+
+    public void delete(Long id, Long usuarioId) {
+        Tarefa tarefaExistente = findById(id);
+        if (!tarefaExistente.getUsuario_criado().getId().equals(usuarioId)) {
+            throw new BadRequestException("Você não tem permissão para deletar esta tarefa.");
+        }
         tarefaRepository.deleteById(id);
+    }
+
+    public int importFromCsv(MultipartFile file) {
+        int count = 0;
+
+        try (BufferedReader br = new BufferedReader(new InputStreamReader(file.getInputStream()))) {
+            String line;
+
+            while ((line = br.readLine()) != null) {
+
+                String[] dados = line.split(",");
+
+                if (dados.length >= 3) {
+                    Long usuarioId = Long.parseLong(dados[2].trim());
+
+                    TarefaDto dto = new TarefaDto(
+                            dados[0].trim(),
+                            dados[1].trim(),
+                            usuarioId
+                    );
+
+
+                    create(dto);
+                    count++;
+                }
+            }
+        } catch (Exception e) {
+            throw new RuntimeException("Erro ao processar CSV: " + e.getMessage());
+        }
+
+        return count;
     }
 
 }
